requirements.txt
behave==1.2.6
selenium==4.0.0
PyHamcrest>=2.0.2
webdriver-manager==4.0.1
allure-python-commons==2.13.2
allure-behave==2.13.2
